Add files via upload
This is a gofish game that pits a computer against you the user. the rules for the game are as followed.
Seven cards are dealt from a standard 52-card deck to each player.  The remaining cards in the deck are shared between the players, usually spread out in a disorderly pile referred to as the "ocean" or "pool". The player whose turn it is to play asks another player for his or her cards of a particular rank.   For example, Alice may ask, "Bob, do you have any threes?" Alice must have at least one card of the rank she requested.  Bob must hand over all cards of that rank if possible. If he has none, Bob tells Alice to "go fish" (or just simply "fish"), and Alice draws a card from the pool and places it in her own hand. Then it is the next player's turn � unless the card Alice drew is the card she asked for, in which case she shows it to the other players, and she gets another turn. When any player at any time has all four cards of one face value, it forms a "book," and the cards must be placed face up in front of that player.  The player that forms the book then gets another turn.  The first person to get 7 "books" wins the game!